requirements.txt
# requirements.txt.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==3.0.1
Jinja2==3.1.3
Pillow>=11.0.0
wheel
gunicorn>=23.0.0
python-dotenv>=1.0.0
psycopg2-binary